Initial update to 1.21.10 (#2304)
The server side changes are pretty straightforward:

 - Level.isClientSide is now a method.
 - NeoForge has switched to a Fabric-style item/fluid storage API. This
   is a fairly simple migration, as we've got support for this already!

   This does mean that we need a wrapper type here too. We really should
   support generics in the generic method system :/.

Unfortunately, this is another rough update on the rendering front. The
main change here is that BE renderers now a) have separate extract/render
phases and b) "draw" to a scene graph (SubmitNodeCollector) rather than
directly.

 - Update lectern models to use the model/layer system, rather than
   constructing model parts directly. We don't *need* to do this, but
   makes this a bit more consistent with vanilla.

 - Remove TurtleUpgradeModel.renderForLevel. I could not figure out a way
   to make this look nice with the separate extract/render phases. This
   is made worse by the fact that CC:T doesn't actually use this
   anywhere, so not sure what we actually need! Something to bring back
   in the future, if other people need it.

 - The turtle/monitor F3 info now uses vanilla's debug renderer system.
   We could maybe make the monitor one use a gizmo in future updates, so
   it renders coordinates in-world instead. Not sure.

Unfortunately, some bits are entirely broken right now:

 - Breaking progress for turtles. There's no way to do custom breaking
   progress, so we just render this as normal custom geometry, which has
   all sorts of depth-buffer issues.

 - Monitors do not use a VBO (again, no way to do this), and render every
   frame. This means monitors are pretty slow to render right now (one or
   two is fine, but monitor arrays are unusable). Will need to submit PRs
   to the various mod loaders for this.

+++ projects/common/src/main/java/dan200/computercraft/shared/peripheral/generic/methods/AbstractEnergyMethods.java
@@ -40,7 +40,7 @@ public abstract class AbstractEnergyMethods<T> implements GenericPeripheral {
      * @return The energy stored in this block, in FE.
      */
     @LuaFunction(mainThread = true)
-    public abstract int getEnergy(T energy);
+    public abstract long getEnergy(T energy);
 
     /**
      * Get the maximum amount of energy this block can store.
@@ -49,5 +49,5 @@ public abstract class AbstractEnergyMethods<T> implements GenericPeripheral {
      * @return The energy capacity of this block.
      */
     @LuaFunction(mainThread = true)
-    public abstract int getEnergyCapacity(T energy);
+    public abstract long getEnergyCapacity(T energy);
 }
